Add errorhandling and QoL
- Add errorhandling for dupes
- Add overwrite possibility for tiles
- Didn't write unittests

+++ main.py
@@ -6,6 +6,7 @@ from datetime import datetime
 import json
 import requests
 from local_secrets import DISCORD_API_KEY
+from errors import TileExistsError
 
 intents = discord.Intents.default()
 intents.message_content = True
@@ -15,6 +16,21 @@ client = commands.Bot(command_prefix='!', intents=intents)
 async def on_ready():
     print('Bot ready')
 
+
+@client.command(pass_context = True)
+async def bhelp(ctx):
+
+    embed = discord.Embed(
+        colour = discord.Colour.orange()
+    )
+
+    embed.set_author(name='Help')
+    embed.add_field(name='!register', value='Register a bingo player (Example. !register Elf)', inline=False)
+    embed.add_field(name='!submit', value='Submit an entry for a tile (Example. !submit 13 Elf)', inline=False)
+    embed.add_field(name='!get_all_entries', value='Get all entries for user (Example. !get_all_entries Elf)', inline=False)
+    embed.add_field(name='!get_entry', value='Register a bingo player (Example. !get_entry 13 Elf)', inline=False)
+
+    await ctx.send(embed=embed)
 
 @client.command(pass_context=True)
 @commands.has_role('Bingo Master')
@@ -59,6 +75,13 @@ async def register(ctx, *args):
 async def submit(ctx, tile, *args):
     # Get the name from the args (can contain spaces) and form a path
     name = " ".join(args)
+    overwrite = False
+
+    if "--ow" in name:
+        split_name = name.split("--ow")
+        name = split_name[0].strip()
+        overwrite = True
+
     path = os.path.dirname(__file__) + '/' + name
     file_exists = os.path.isdir(path)
     
@@ -67,7 +90,12 @@ async def submit(ctx, tile, *args):
         return
         
     # If the account exists, create a text entry of the submission
-    create_submit_entry(path, tile)
+    print(overwrite)
+    try:
+        create_submit_entry(path, tile, overwrite)
+    except TileExistsError as e:
+        await ctx.send(e)
+        return
 
     # If the account exists, create an image entry of the submission
     img_data = requests.get(ctx.message.attachments[0].url).content
@@ -78,7 +106,7 @@ async def submit(ctx, tile, *args):
     await ctx.send('Submission saved for account %s, Tile: %s.' % (name, tile))
 
 
-def create_submit_entry(path, tile):
+def create_submit_entry(path, tile, overwrite=False):
     path = path + '/entries.json'
     file_exists = os.path.isfile(path)
 
@@ -88,10 +116,27 @@ def create_submit_entry(path, tile):
         with open(path, 'r') as json_file:
             data = json.load(json_file)
 
-        data['entries'].append({
-            'tile': tile,
-            'submitted': datetime.now().strftime("%d/%m/%Y, %H:%M:%S")
-        })
+        tile_exists = False
+        found_tile_index = -1
+
+        for index, entry in enumerate(data['entries']):
+            if entry['tile'] == tile:
+                tile_exists = True
+                found_tile_index = index
+                break
+
+        if not overwrite and tile_exists:
+            raise TileExistsError("Tile already exists for that id. If you want to overwrite the data, use --ow (i.e. !submit 2 Elf --ow)")
+        
+        if not tile_exists:
+            data['entries'].append({
+                'tile': tile,
+                'submitted': datetime.now().strftime("%d/%m/%Y, %H:%M:%S")
+            })
+        else:
+            data['entries'][found_tile_index]['tile'] = tile
+            data['entries'][found_tile_index]['submitted'] = datetime.now().strftime("%d/%m/%Y, %H:%M:%S")
+
         with open(path, 'w') as json_file:
             json_string = json.dumps(data)
             json_file.write(json_string)
